backend/requirements.txt
fastapi
uvicorn
redis
aioboto3
pyjwt
python-multipart
prometheus-fastapi-instrumentator
